fix: move model fetch api into the top level component
